ui: move LabelWithHelp in common packages
