+++ ui/src/authors/components/CitationSummaryTable/CitationSummaryTable.scss
@@ -8,7 +8,10 @@
   th {
     color: $secondary-color;
     font-weight: normal;
-    padding: 1rem;
+    padding-left: 0.5rem;
+    padding-right: 0.25rem;
+    padding-top: 0.75rem;
+    padding-bottom: 0.75rem;
   }
 
   td {
